Set Pomodoro timer from wall clock rather than counting down
The current `org-pomodoro-tick` function, which decrements
`org-pomodoro-countdown` once a second for the number of seconds in the
Pomodoro, is not robust if Emacs fails to run for more than a second at a
time. The most common scenarios where this happens are

- putting the computer to sleep (most relevant on laptops)
- suspending Emacs in the terminal using C-z or `SIGSTOP`.

To work around this, calculate the end time of the pomodoro phase upon starting
it, and check in `org-pomodoro-tick` whether the current time is past the end
time to determine when to end the phase.

Fix[ https://github.com/lolownia/org-pomodoro/issues/66.